Fix crashing on stores display. I wasn't very smart when I defined store's balance as int instead of double.

+++ lib/models/store.dart
@@ -1,7 +1,7 @@
 class Store {
   late int storeId;
   late String storeName;
-  late int amount;
+  late double amount;
   late int employeeCount;
 
   Store(this.storeId, this.storeName, this.amount, this.employeeCount);
